// src/interfaces.ts
export interface ParsedAccount extends AccountMeta {
	/** Account name, same as in Idl, nested accounts look like `account.nestedAccount` */
	name?: string;
}

/**
 * Instructions args with correct types for specific instruction by instruction name
 */
export type ParsedIdlArgsByInstructionName<I extends Idl, Ix extends I["instructions"][number]> = {
	[ArgName in Ix["args"][number]["name"]]: DecodeType<(Ix["args"][number] & { name: ArgName })["type"], IdlTypes<I>>;
};

export type InstructionNames<I extends Idl> = I["instructions"][number]["name"];

export type ParsedIdlArgs<I extends Idl, IxName extends InstructionNames<I> = InstructionNames<I>> = ParsedIdlArgsByInstructionName<I, IxByName<I, IxName>>;

export type ParsedArgs = {
	[key: string]: unknown;
};

export type UnknownInstruction = {
	name: "unknown" | string;
	args: { unknown: unknown };
	accounts: ParsedAccount[];
	programId: PublicKey;
};

export type ParsedInstruction<I extends Idl, IxName extends InstructionNames<I> = InstructionNames<I>> =
	| UnknownInstruction
	| ParsedIdlInstruction<I, IxName>
	| ParsedCustomInstruction;

export interface ParsedCustomInstruction {
	/** Instruction name */
	name: string;
	programId: PublicKey;
	/** Parsed arguments */
	args: unknown;
	/** Parsed accounts */
	accounts: ParsedAccount[];
}

export interface ParsedIdlInstruction<I extends Idl, IxName extends InstructionNames<I> = InstructionNames<I>> {
	/** Instruction name */
	name: IxName;
	programId: PublicKey;
	/** Parsed arguments */
	args: ParsedIdlArgs<I, IxName>;
	/** Parsed accounts */
	accounts: IdlAccountsToFlatMeta<IxByName<I, IxName>["accounts"]>;
}

export type IdlInstructionAccountItem2 = IdlInstructionAccount | IdlInstructionAccounts;
export type IdlInstructionAccount = {
	name: string;
	docs?: string[];
	writable?: boolean;
	signer?: boolean;
	optional?: boolean;
	address?: string;
	relations?: string[];
};
export type IdlInstructionAccounts = {
	name: string;
	accounts: IdlInstructionAccountItem2[];
};

export type IdlAccountsToFlatMeta<T extends IdlInstructionAccountItem2[], Prefix extends string = ""> = T extends [infer First, ...infer Rest]
	? First extends IdlInstructionAccounts
		? [
				...IdlAccountsToFlatMeta<First["accounts"], `${Prefix}${First["name"]}.`>,
				...IdlAccountsToFlatMeta<Rest extends IdlInstructionAccountItem2[] ? Rest : [], Prefix>,
			]
		: First extends IdlInstructionAccount
			? [
					{
						name: `${Prefix}${First["name"]}`;
						isSigner: First["signer"] extends boolean ? First["signer"] : false;
						isWritable: First["writable"] extends boolean ? First["writable"] : false;
						pubkey: PublicKey;
					},
					...IdlAccountsToFlatMeta<Rest extends IdlInstructionAccountItem2[] ? Rest : [], Prefix>,
				]
			: never
	: [];
// ...
export type IxByName<I extends Idl, IxName extends I["instructions"][number]["name"]> = I["instructions"][number] & { name: IxName };